+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## [0.0.2] - 2024-11-09
+### Added
+- Added the command cycleAllQuickFixes which cycles through all areas with a quick fix one by one
+- Added a new setting to turn off the behavior to automatically open the quick fix menu when clicking on an area with an available quick fix
+
 ## [0.0.1] - 2024-11-08
 ### Initial release